Moves files to better more appropriate folders
