+++ index.html
@@ -1,6 +1,6 @@
 <html>
-    <a href="experiment.html">Experiment</a>
-    <a href="serial-test.html">Testing serial port</a>
-    <a href="stars-visualiser.html">Testing stars</a>
-    <a href="three-index.html">Three path integration</a>
+    <a href="public/experiment.html">Experiment</a>
+    <a href="public/serial-test.html">Testing serial port</a>
+    <a href="public/stars-visualiser.html">Testing stars</a>
+    <a href="public/three-index.html">Three path integration</a>
 </html>
