+++ public/experiment.html
@@ -1,9 +1,9 @@
 <html>
     <head>
-        <script src="src/css/js/jquery.min.js"></script>
-        <script src="src/css/js/bootstrap.bundle.min.js"></script>
-        <link rel="stylesheet" href="src/css/bootstrap.min.css">
-        <link rel="stylesheet" href="src/css/style.css">
+        <script src="css/js/jquery.min.js"></script>
+        <script src="css/js/bootstrap.bundle.min.js"></script>
+        <link rel="stylesheet" href="css/bootstrap.min.css">
+        <link rel="stylesheet" href="css/style.css">
     </head>
     <body>
         <div id="setup">
@@ -50,6 +50,6 @@
                 <a href="#" id="btn-back" class="star-button" style="left:30%">Back</a>
             </div>
         </div>
-        <script type="module" src='src/pages/app.js'></script>
+        <script type="module" src='../src/pages/app.js'></script>
     </body>
 </html>
