+++ public/serial-test.html
@@ -27,7 +27,7 @@
   </div>
   <div id="serial-messages-container"></div>
 
-  <script src="src/serial.js"></script>
+  <script src="../src/serial.js"></script>
   <script>
     const arduinoController = new ArduinoController();
     const getSerialMessages = document.getElementById('get-serial-messages');
