+++ public/stars-visualiser.html
@@ -1,9 +1,9 @@
 <html>
     <head>
-        <script src="src/css/js/jquery.min.js"></script>
-        <script src="src/css/js/bootstrap.bundle.min.js"></script>
-        <link rel="stylesheet" href="src/css/bootstrap.min.css">
-        <link rel="stylesheet" href="style.css">
+        <script src="css/js/jquery.min.js"></script>
+        <script src="css/js/bootstrap.bundle.min.js"></script>
+        <link rel="stylesheet" href="css/bootstrap.min.css">
+        <link rel="stylesheet" href="css/style.css">
     </head>
     <body>
         <div id="setup">
@@ -32,7 +32,7 @@
                 <canvas id="space"></canvas>
             </div>
         </div>
-        <script src="src/serial.js"></script>
-        <script type="module" src='src/pages/stars-visualiser.js'></script>
+        <script src="../src/serial.js"></script>
+        <script type="module" src='../src/pages/stars-visualiser.js'></script>
     </body>
 </html>
