+++ public/three-index.html
@@ -3,7 +3,6 @@
 		<title>three.js webgl - geometry - minecraft</title>
 		<meta charset="utf-8">
 		<meta name="viewport" content="width=device-width, user-scalable=no, minimum-scale=1.0, maximum-scale=1.0">
-        <link type="text/css" rel="stylesheet" href="./three/examples/main.css">
         <style>
 			#blocker {
 				position: absolute;
@@ -62,13 +61,13 @@
         </div>
 		<script type="module">
 
-			import * as THREE from './src/three/three.module.js';
-            import getRandomStarField from './src/three/starfield.js';
-			import Stats from './src/three/stats.module.js';
+			import * as THREE from '../src/three/three.module.js';
+            import getRandomStarField from '../src/three/starfield.js';
+			import Stats from '../src/three/stats.module.js';
 
-            import { BufferGeometryUtils } from './src/three/BufferGeometryUtils.js';
+            import { BufferGeometryUtils } from '../src/three/BufferGeometryUtils.js';
             
-            import { GUI } from './src/three/dat.gui.module.js';
+            import { GUI } from '../src/three/dat.gui.module.js';
 
             class Trial{
                 constructor(start, goal1, goal2, decoy){
